♻️  rename boot/[package].ts to boot/start[package].ts
