+++ packages/logs/src/boot/startLogs.ts
