+++ packages/logs/src/index.ts
@@ -1,4 +1,4 @@
 export { Logger, LogsMessage, StatusType, HandlerType } from './domain/logger'
 export { LoggerConfiguration, LogsPublicApi as LogsGlobal, datadogLogs } from './boot/logs.entry'
-export { LogsUserConfiguration } from './boot/logs'
+export { LogsUserConfiguration } from './boot/startLogs'
 export { LogsEvent } from './logsEvent.types'
