+++ packages/rum-core/src/boot/startRum.ts
