+++ packages/rum-core/src/index.ts
@@ -10,7 +10,7 @@ export {
   RumLongTaskEvent,
 } from './rumEvent.types'
 export { ViewContext, CommonContext } from './rawRumEvent.types'
-export { startRum } from './boot/rum'
+export { startRum } from './boot/startRum'
 export { LifeCycle, LifeCycleEventType } from './domain/lifeCycle'
 export { ParentContexts } from './domain/parentContexts'
 export { RumSession } from './domain/rumSession'
