+++ packages/rum-recorder/src/boot/recorder.entry.ts
@@ -1,7 +1,7 @@
 import { defineGlobal, getGlobalObject } from '@datadog/browser-core'
 import { startRum } from '@datadog/browser-rum-core'
 
-import { startRecording } from './recorder'
+import { startRecording } from './startRecording'
 import { RumRecorderPublicApi, makeRumRecorderPublicApi } from './rumRecorderPublicApi'
 
 export const datadogRum = makeRumRecorderPublicApi(startRum, startRecording)
