+++ packages/rum-recorder/src/boot/startRecording.ts
